+++ SpeedDate.ClientPlugins.Peer/Echo/EchoPlugin.cs
@@ -8,7 +8,7 @@ namespace SpeedDate.ClientPlugins.Peer.Echo
     {
         public void Send(string message, Action<string> echoCallback, ErrorCallback error)
         {
-            Client.SendMessage((ushort)OpCodes.Echo, message, (status, response) => 
+            Client.SendMessage((uint)OpCodes.Echo, message, (status, response) => 
             { 
                 if (status != ResponseStatus.Success)
                 {
